+++ power-management.h
@@ -2,7 +2,7 @@
 #define _POWER_MANAGEMENT_H_
 #include <power-management.h>
 
-extern uint8_t power_management_inactivity;
+extern uint16_t power_management_inactivity;
 
 #define power_management_pwr_down_reset() power_management_inactivity = 0
 #define power_management_pwr_down_set_idle() power_management_inactivity++
@@ -10,7 +10,7 @@ extern uint8_t power_management_inactivity;
 	power_management_inactivity = value
 
 void
-power_management_power_down_init(uint8_t inactivity_timeout,
+power_management_power_down_init(uint16_t inactivity_timeout,
 				 void (*on_sleep)(void *arg), void *arg);
 
 #endif
